#68 rename: global 패키지 재분류 및 RestTemplateConfig 제거
