+++ src/main/java/com/carumuch/capstone/domain/auth/service/AuthService.java
@@ -4,8 +4,8 @@ import com.carumuch.capstone.domain.auth.dto.TokenDto;
 import com.carumuch.capstone.domain.auth.dto.VerificationCodeDto;
 import com.carumuch.capstone.domain.auth.dto.VerificationLoginIdDto;
 import com.carumuch.capstone.domain.auth.jwt.JwtTokenProvider;
-import com.carumuch.capstone.global.common.ErrorCode;
-import com.carumuch.capstone.global.common.exception.CustomException;
+import com.carumuch.capstone.global.exception.ErrorCode;
+import com.carumuch.capstone.global.exception.CustomException;
 import com.carumuch.capstone.domain.user.model.User;
 import com.carumuch.capstone.domain.user.repository.UserRepository;
 import com.carumuch.capstone.global.service.MailService;
