+++ src/main/java/com/carumuch/capstone/domain/board/model/Board.java
@@ -1,7 +1,7 @@
 package com.carumuch.capstone.domain.board.model;
 
 import com.carumuch.capstone.domain.comment.model.Comment;
-import com.carumuch.capstone.global.auditing.BaseCreateByEntity;
+import com.carumuch.capstone.global.base.BaseCreateByEntity;
 import com.carumuch.capstone.domain.user.model.User;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
