+++ src/main/java/com/carumuch/capstone/domain/board/service/BoardService.java
@@ -6,8 +6,8 @@ import com.carumuch.capstone.domain.board.dto.BoardModifyReqDto;
 import com.carumuch.capstone.domain.board.dto.BoardReqDto;
 import com.carumuch.capstone.domain.board.repository.BoardImageRepository;
 import com.carumuch.capstone.domain.board.repository.BoardRepository;
-import com.carumuch.capstone.global.common.ErrorCode;
-import com.carumuch.capstone.global.common.exception.CustomException;
+import com.carumuch.capstone.global.exception.ErrorCode;
+import com.carumuch.capstone.global.exception.CustomException;
 import com.carumuch.capstone.domain.image.service.ImageService;
 import com.carumuch.capstone.domain.user.model.User;
 import com.carumuch.capstone.domain.user.repository.UserRepository;
