+++ src/main/java/com/carumuch/capstone/domain/bodyshop/service/BodyShopService.java
@@ -13,8 +13,8 @@ import com.carumuch.capstone.domain.bodyshop.dto.BodyShopRegistrationReqDto;
 import com.carumuch.capstone.domain.bodyshop.dto.BodyShopUpdateReqDto;
 import com.carumuch.capstone.domain.bid.repository.BidRepository;
 import com.carumuch.capstone.domain.bodyshop.repository.BodyShopRepository;
-import com.carumuch.capstone.global.common.ErrorCode;
-import com.carumuch.capstone.global.common.exception.CustomException;
+import com.carumuch.capstone.global.exception.ErrorCode;
+import com.carumuch.capstone.global.exception.CustomException;
 import com.carumuch.capstone.domain.user.model.User;
 import com.carumuch.capstone.domain.user.repository.UserRepository;
 import com.carumuch.capstone.domain.estimate.model.Estimate;
