+++ src/main/java/com/carumuch/capstone/domain/comment/service/CommentService.java
@@ -6,8 +6,8 @@ import com.carumuch.capstone.domain.comment.model.Comment;
 import com.carumuch.capstone.domain.comment.dto.CommentModifyReqDto;
 import com.carumuch.capstone.domain.comment.dto.CommentReqDto;
 import com.carumuch.capstone.domain.comment.repository.CommentRepository;
-import com.carumuch.capstone.global.common.ErrorCode;
-import com.carumuch.capstone.global.common.exception.CustomException;
+import com.carumuch.capstone.global.exception.ErrorCode;
+import com.carumuch.capstone.global.exception.CustomException;
 import com.carumuch.capstone.domain.user.model.User;
 import com.carumuch.capstone.domain.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
