+++ src/main/java/com/carumuch/capstone/domain/estimate/service/EstimateService.java
@@ -4,8 +4,8 @@ import com.carumuch.capstone.domain.bid.model.type.BidStatus;
 import com.carumuch.capstone.domain.bid.repository.BidRepository;
 import com.carumuch.capstone.domain.estimate.dto.*;
 import com.carumuch.capstone.domain.estimate.repository.EstimateRepository;
-import com.carumuch.capstone.global.common.ErrorCode;
-import com.carumuch.capstone.global.common.exception.CustomException;
+import com.carumuch.capstone.global.exception.ErrorCode;
+import com.carumuch.capstone.global.exception.CustomException;
 import com.carumuch.capstone.domain.user.model.User;
 import com.carumuch.capstone.domain.user.repository.UserRepository;
 import com.carumuch.capstone.domain.estimate.model.Estimate;
@@ -21,7 +21,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import static com.carumuch.capstone.global.common.ErrorCode.BID_ALREADY_COMPLETED;
+import static com.carumuch.capstone.global.exception.ErrorCode.BID_ALREADY_COMPLETED;
 
 
 @Service
