+++ src/main/java/com/carumuch/capstone/domain/user/service/UserService.java
@@ -7,8 +7,8 @@ import com.carumuch.capstone.domain.user.dto.UserUpdateReqDto;
 import com.carumuch.capstone.domain.user.model.User;
 import com.carumuch.capstone.domain.user.model.type.Role;
 import com.carumuch.capstone.domain.user.repository.UserRepository;
-import com.carumuch.capstone.global.common.ErrorCode;
-import com.carumuch.capstone.global.common.exception.CustomException;
+import com.carumuch.capstone.global.exception.ErrorCode;
+import com.carumuch.capstone.global.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.context.SecurityContextHolder;
