+++ src/main/java/com/carumuch/capstone/global/config/SecurityConfig.java
@@ -6,9 +6,9 @@ import com.carumuch.capstone.domain.auth.filter.JwtAuthenticationFilter;
 import com.carumuch.capstone.domain.auth.jwt.*;
 import com.carumuch.capstone.domain.auth.service.AuthService;
 import com.carumuch.capstone.domain.auth.service.CustomOAuth2UserService;
-import com.carumuch.capstone.global.common.exception.CustomAccessDeniedHandler;
-import com.carumuch.capstone.global.common.exception.CustomAuthenticationEntryPoint;
-import com.carumuch.capstone.global.common.exception.CustomOAuth2AuthenticationFailureHandler;
+import com.carumuch.capstone.domain.auth.handler.CustomAccessDeniedHandler;
+import com.carumuch.capstone.domain.auth.handler.CustomAuthenticationEntryPoint;
+import com.carumuch.capstone.domain.auth.handler.CustomOAuth2AuthenticationFailureHandler;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Value;
